@@ -1,66 +1,69 @@
 {
-  "name": "SmartMonitor",
+  "name": "smart_monitor",
   "version": "1.0.0",
-  "description": "Smart Monitor Application",
-  "main": "index.js",
+  "description": "Smart Monitor: A Digital Signage MQTT IoT Based Solution Server",
+  "main": "./dist/server.js",
   "scripts": {
     "build": "tsc",
     "dev": "ts-node ./lib/server.ts",
-    "start": "nodemon ./dist/server.js",
+    "start": "nodemon -e js ./dist/server.js",
     "deploy": "node ./dist/server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
+  "author": "Pratik Chowdhury (pratikc@live.co.uk)",
   "license": "BSD-2-Clause",
   "dependencies": {
-    "@types/ejs": "^3.0.2",
+    "@types/ejs": "^3.0.4",
     "@types/express-session": "^1.17.0",
     "@types/fluent-ffmpeg": "^2.1.14",
-    "@types/mime": "^2.0.1",
-    "@types/mongodb": "^3.5.5",
-    "@types/multer": "^1.4.2",
-    "@types/passport": "^1.0.3",
+    "@types/mime": "^2.0.3",
+    "@types/mongodb": "^3.5.25",
+    "@types/multer": "^1.4.3",
+    "@types/passport": "^1.0.4",
     "@types/passport-local": "^1.0.33",
-    "@types/pg": "^7.14.3",
+    "@types/pg": "^7.14.4",
     "@types/session-file-store": "^1.2.1",
-    "@types/sharp": "^0.24.0",
-    "bcrypt": "^4.0.1",
+    "@types/sharp": "^0.25.1",
+    "@types/validator": "^13.1.0",
+    "bcrypt": "^5.0.0",
     "body-parser": "^1.19.0",
-    "connect-session-sequelize": "^6.1.1",
+    "connect-session-sequelize": "^7.0.1",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
-    "ejs": "^3.0.2",
+    "ejs": "^3.1.3",
     "express": "^4.17.1",
-    "express-session": "^1.17.0",
-    "express-validator": "^6.4.0",
+    "express-session": "^1.17.1",
+    "express-validator": "^6.6.0",
     "fluent-ffmpeg": "^2.1.2",
     "gif-info": "^1.0.1",
     "jsonwebtoken": "^8.5.1",
-    "mime": "^2.4.4",
-    "mongodb": "^3.5.5",
-    "mqtt": "^3.0.0",
+    "mime": "^2.4.6",
+    "mongodb": "^3.5.9",
+    "mqtt": "^4.1.0",
     "multer": "^1.4.2",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
-    "pg": "^8.0.2",
+    "pg": "^8.3.0",
     "reflect-metadata": "^0.1.13",
-    "sequelize": "^5.21.6",
-    "sequelize-typescript": "^1.1.0",
+    "sequelize": "^6.1.1",
+    "sequelize-typescript": "next",
     "session-file-store": "^1.4.0",
-    "sharp": "^0.25.2"
+    "sharp": "^0.25.4"
   },
   "devDependencies": {
     "@types/bcrypt": "^3.0.0",
     "@types/body-parser": "^1.19.0",
     "@types/cookie-parser": "^1.4.2",
     "@types/cors": "^2.8.6",
-    "@types/express": "^4.17.6",
-    "@types/jsonwebtoken": "^8.3.9",
-    "@types/node": "^13.11.1",
-    "@types/sequelize": "^4.28.8",
-    "npm-check-updates": "^4.1.2",
-    "nodemon": "^2.0.3",
-    "typescript": "^3.8.3"
+    "@types/express": "^4.17.7",
+    "@types/jsonwebtoken": "^8.5.0",
+    "@types/node": "^14.0.26",
+    "nodemon": "^2.0.4",
+    "npm-check-updates": "^7.0.2",
+    "typescript": "^3.9.7"
+  },
+  "engines": {
+    "node": ">=10.0.0"
   }
 }
