@@ -1,21 +1,18 @@
-import * as dotenv from "dotenv"
+import dotenv from "dotenv";
 
 // If No import taking place from .env file
 // Use this to ensure that dotenv file is loaded
-if (!process.env.Dialect) {
-  dotenv.config();
-}
+dotenv.config();
 
 import express from "express";
-import * as bodyParser from "body-parser";
+import bodyParser from "body-parser";
 import * as Models from "./Models/Sequelize";
-import * as Routes from "./routes/Routes";
-import * as cors from "cors";
+import * as Routes from "./routes";
+import cors from "cors";
 import { PassportModelsGenerate } from "./Models/Passport.Models";
 import { RoutesCommon } from "./routes/Common.Routes";
 
-export async function App() {
-
+export default async function App() {
   const app = express();
 
   // app.use(bodyParser.json());
@@ -25,22 +22,21 @@ export async function App() {
   // Use this to Render HTML
 
   // Specify Path of Website Static Contents
-  app.engine('.html', require('ejs').renderFile);
+  app.engine(".html", require("ejs").renderFile);
   app.set("views", "./Views");
-  app.set('view engine', 'ejs');
+  app.set("view engine", "ejs");
 
   // Only Enable if our app is not dockerised
   // In Dockerised App, static files are served by
   // NGINX
   // Thus reducing the load on server
-  if (!process.env.APP_IS_DOCKERISED) {
+  if (!Boolean(process.env.APP_IS_DOCKERISED)) {
     app.use("/static/", express.static("./Website"));
   }
-
   await Models.RunSynchronisation();
   await PassportModelsGenerate(app);
 
-  app.use(cors.default());
+  app.use(cors());
   // middleware for json body parsing
   app.use(bodyParser.json());
   // middleware for parsing application/x-www-form-urlencoded
@@ -58,8 +54,7 @@ export async function App() {
   app.get("/navbar", RoutesCommon.IsAuthenticated, (req, res) => {
     if (RoutesCommon.GetUser(req).Authority === "ADMIN")
       return res.render("navbaradmin.html");
-    else
-      return res.render("navbarnormal.html");
+    else return res.render("navbarnormal.html");
   });
 
   app.get("/", (req, res) => {
