@@ -1,16 +1,18 @@
-import { IfDockerisedOrSelectDefault } from "./Config.Common";
+export const DatabaseName = String(
+  process.env.MongoDatabaseName || "smartmonitor"
+);
+export const Host = String(process.env.MongoHost || "localhost");
+export const Port = Number(process.env.MongoPort || 27017);
+export const UserName = String(
+  process.env.MONGO_INITDB_ROOT_USERNAME || "mongo"
+);
+export const Password = String(
+  process.env.MONGO_INITDB_ROOT_PASSWORD || "mongo"
+);
 
-export namespace Mongo {
-    export const DatabaseName = String(process.env.MongoDatabaseName);
-    export const Host = IfDockerisedOrSelectDefault(process.env.MongoHost, "localhost");
-    export const Port = Number(process.env.MongoPort);
-    export const UserName = String(process.env.MONGO_INITDB_ROOT_USERNAME);
-    export const Password = String(process.env.MONGO_INITDB_ROOT_PASSWORD);
+export const Bucket = String(process.env.MongoBucketName || "smartmonitor");
+export const Bucket_Files = Bucket + ".files";
 
-    export const Bucket = String(process.env.MongoBucketName);
-    export const Bucket_Files = Mongo.Bucket + ".files";
-
-    export function Uri() {
-        return "mongodb://" + Mongo.UserName + ":" + Mongo.Password + "@" + Mongo.Host + ":" + Mongo.Port;
-    }
-};
+export function Uri() {
+  return "mongodb://" + UserName + ":" + Password + "@" + Host + ":" + Port;
+}
