@@ -1,15 +1,13 @@
-import { GetIP } from "./Config.Common";
+// import { GetIP } from "./Config.Common";
 
 function GetMqttAddress() {
-  if (process.env.MQTT_URI_CONNECTOR) {
+  if (process.env.MQTT_URI_CONNECTOR != null) {
     return "tcp://" + process.env.MQTT_URI_CONNECTOR + ":1883/";
   }
-  return "tcp://" + GetIP() + ":1883/";
+  return "tcp://" + "localhost" + ":1883/";
 }
 
-export const Mqtt = {
-  Url: GetMqttAddress(),
-  DisplayTopic: (id: number) => {
-    return "/display/" + id;
-  }
-};
+export const Url = GetMqttAddress();
+export function DisplayTopic(id: number) {
+  return "/display/" + id;
+}
