@@ -1,8 +1,8 @@
-import {GetIP} from "./Config.Common";
-import * as Path from "path";
+import Path from "path";
 
-export const Server = {
-  Port: 8000,
-  Name: GetIP(),
-  MediaStorage: Path.resolve(String(process.env.INIT_CWD), 'uploads')
-};
+export const Port = 8000;
+export const Name = "0.0.0.0";
+export const MediaStorage = Path.resolve(
+  String(process.env.INIT_CWD),
+  "uploads"
+);
