@@ -1,12 +1,12 @@
 export const Session = {
-  secret: "SmartMonitor",
+  secret: String(process.env.SESSION_SECRET || "SmartMonitor"),
   cookie: {
     // Max Time for Cookie Existence
     // In MilliSeconds
     maxAge: 4 * 24 * 60 * 60 * 1000,
-    secure: false
+    secure: false,
   },
   name: "Smart Monitor",
   resave: false,
-  saveUninitialized: true
+  saveUninitialized: true,
 };
