@@ -1,12 +1,18 @@
 import { Router } from "express";
 import { RoutesCommon, upload } from "./Common.Routes";
-import * as Models from "../Models/Models";
-import * as Path from "path";
-import * as Config from "../config/Config";
+import * as Models from "../Models";
+import Path from "path";
+import * as Config from "../config";
 
 export const Files = Router();
 
-async function NewFileAtPathAdded(path: string, displayIDs: number[], showTime: number, startTime: number, endTime: number) {
+async function NewFileAtPathAdded(
+  path: string,
+  displayIDs: number[],
+  showTime: number,
+  startTime: number,
+  endTime: number
+) {
   let extension = Path.extname(path).substr(1).toLowerCase(); // Ignores Dot
 
   let mediaType = RoutesCommon.GetFileMediaType(extension);
@@ -31,7 +37,7 @@ async function NewFileAtPathAdded(path: string, displayIDs: number[], showTime: 
   const fileHash = await RoutesCommon.GetSHA256FromFileAsync(path);
 
   const fileSame = await Models.Files.findOne({
-    where: { FileHash: fileHash }
+    where: { FileHash: fileHash },
   });
 
   const AlreadyPresentIDs: number[] = [];
@@ -41,8 +47,7 @@ async function NewFileAtPathAdded(path: string, displayIDs: number[], showTime: 
     mediaType = fileSame.MediaType;
 
     AlreadyPresentIDs.push(fileSame.DisplayID);
-  }
-  else {
+  } else {
     // As it's New File
     // Generate Thumbnail
     await RoutesCommon.GenerateThumbnailAsync(
@@ -50,7 +55,10 @@ async function NewFileAtPathAdded(path: string, displayIDs: number[], showTime: 
       mediaType,
       Models.Files.GetThumbnailFileName(path)
     );
-    await Models.Mongo.File.UploadSingle(Path.join(location, Models.Files.GetThumbnailFileName(name)), Models.Files.GetThumbnailFileName(fileHash));
+    await Models.Mongo.File.UploadSingle(
+      Path.join(location, Models.Files.GetThumbnailFileName(name)),
+      Models.Files.GetThumbnailFileName(fileHash)
+    );
 
     if (extension === "gif") {
       showTime = await RoutesCommon.GIFDuration(path, showTime);
@@ -69,20 +77,19 @@ async function NewFileAtPathAdded(path: string, displayIDs: number[], showTime: 
           OnDisplay: true,
           TimeStart: startTime,
           TimeEnd: endTime,
-          ShowTime: showTime
+          ShowTime: showTime,
         },
         {
           where: {
             Extension: extension,
             DisplayID: displayId,
             FileHash: fileHash,
-            MediaType: mediaType!
-          }
+            MediaType: mediaType!,
+          },
         }
       );
       displaysWhereUpdated.push(displayId);
-    }
-    else {
+    } else {
       await Models.Files.create({
         Extension: extension,
         DisplayID: displayId,
@@ -92,7 +99,7 @@ async function NewFileAtPathAdded(path: string, displayIDs: number[], showTime: 
         TimeStart: startTime,
         TimeEnd: endTime,
         ShowTime: showTime,
-        Downloaded: false
+        Downloaded: false,
       });
 
       displaysWhereDownload.push(displayId);
@@ -102,14 +109,29 @@ async function NewFileAtPathAdded(path: string, displayIDs: number[], showTime: 
   if (!fileSame)
     await Models.Mongo.File.UploadSingle(path, fileHash + "." + extension);
 
-  return [displaysWhereUpdated, displaysWhereDownload];
+  return { displaysWhereUpdated, displaysWhereDownload };
 }
-async function NewFilesAtPathAdded(paths: string[], displayIDs: number[], showTime: number, startTime: number, endTime: number) {
-  const promises: Promise<number[][]>[] = [];
+async function NewFilesAtPathAdded(
+  paths: string[],
+  displayIDs: number[],
+  showTime: number,
+  startTime: number,
+  endTime: number
+) {
+  const promises: Promise<{
+    displaysWhereUpdated: number[];
+    displaysWhereDownload: number[];
+  }>[] = [];
   // Iterate over all the files
   for (const path of paths) {
     // Path changed if path is a video
-    const promise = NewFileAtPathAdded(path, displayIDs, showTime, startTime, endTime);
+    const promise = NewFileAtPathAdded(
+      path,
+      displayIDs,
+      showTime,
+      startTime,
+      endTime
+    );
     promises.push(promise);
   }
   let displayWhereDownload: number[] = [];
@@ -118,9 +140,9 @@ async function NewFilesAtPathAdded(paths: string[], displayIDs: number[], showTi
   // Merge all Display Details together
   // So we can send Signals
   const displays = await Promise.all(promises);
-  for (const [updated, download] of displays) {
-    displayWhereDownload.push(...download);
-    displaysWhereUpdated.push(...updated);
+  for (const { displaysWhereUpdated, displaysWhereDownload } of displays) {
+    displayWhereDownload.push(...displaysWhereDownload);
+    displaysWhereUpdated.push(...displaysWhereUpdated);
   }
 
   // Contains all Unique Displays where Deletion and Updation Occured
@@ -128,9 +150,11 @@ async function NewFilesAtPathAdded(paths: string[], displayIDs: number[], showTi
   displaysWhereUpdated = [...new Set(displaysWhereUpdated)];
 
   // Now do not send Update Signal to Displays where Create Signal being sent
-  // This is because On Receiving Create Signal, the action of 
-  // Update Signal is Also Performed 
-  displaysWhereUpdated = displaysWhereUpdated.filter(display => !displayWhereDownload.includes(display));
+  // This is because On Receiving Create Signal, the action of
+  // Update Signal is Also Performed
+  displaysWhereUpdated = displaysWhereUpdated.filter(
+    (display) => !displayWhereDownload.includes(display)
+  );
 
   await RoutesCommon.Mqtt.SendDownloadRequests(displayWhereDownload);
   await RoutesCommon.Mqtt.SendUpdateSignals(displaysWhereUpdated);
@@ -163,8 +187,14 @@ Files.post(
         0 /*Default Time Set*/
       );
 
-      const paths = files.map(value => value.path);
-      await NewFilesAtPathAdded(paths, displayIDs, showTime, startTime, endTime);
+      const paths = files.map((value) => value.path);
+      await NewFilesAtPathAdded(
+        paths,
+        displayIDs,
+        showTime,
+        startTime,
+        endTime
+      );
 
       await RemoveFilesNotInDB();
 
@@ -178,9 +208,11 @@ Files.post(
 
 async function RemoveFilesNotInDB() {
   const FilesInDBObj = await Models.Files.findAll();
-  let FilesInDB = FilesInDBObj.map(file => file.FileName);
+  let FilesInDB = FilesInDBObj.map((file) => file.FileName);
   FilesInDB = [...new Set(FilesInDB)].sort();
-  const FilesInDBThumbnails = FilesInDB.map(file => Models.Files.GetThumbnailFileName(file));
+  const FilesInDBThumbnails = FilesInDB.map((file) =>
+    Models.Files.GetThumbnailFileName(file)
+  );
   FilesInDB.push(...FilesInDBThumbnails);
 
   let FilesInMongo = await Models.Mongo.File.GetAll();
@@ -188,7 +220,9 @@ async function RemoveFilesNotInDB() {
 
   // Only Remove Files not referenced by Database
   // But are present in Mongo
-  const FilesToRemove = FilesInMongo.filter((file) => !FilesInDB.includes(file));
+  const FilesToRemove = FilesInMongo.filter(
+    (file) => !FilesInDB.includes(file)
+  );
 
   await Models.Mongo.File.RemoveMultiple(FilesToRemove);
 }
@@ -199,14 +233,13 @@ Files.delete("/remove", RoutesCommon.IsAuthenticated, async (req, res) => {
   const displayId = Number(params.id);
 
   const FileToDelete = await Models.Files.findOne({
-    where: { id: fileId, DisplayID: displayId }
+    where: { id: fileId, DisplayID: displayId },
   });
 
-  if (FileToDelete == null)
-    return;
+  if (FileToDelete == null) return;
 
   const count = await Models.Files.destroy({
-    where: { id: fileId, DisplayID: displayId }
+    where: { id: fileId, DisplayID: displayId },
   });
 
   if (count === 0) return res.json({ success: false });
@@ -216,7 +249,7 @@ Files.delete("/remove", RoutesCommon.IsAuthenticated, async (req, res) => {
   return res.json({ success: true });
 });
 
-Files.get("/upload/", RoutesCommon.IsAuthenticated, async (req, res) => {
+Files.get("/upload/", RoutesCommon.IsAuthenticated, async (_req, res) => {
   return RoutesCommon.NoCaching(res).render("ImageUpload.html");
 });
 
@@ -231,9 +264,8 @@ Files.put("/shown", RoutesCommon.IsAuthenticated, async (req, res) => {
     const [count] = await Models.Files.update(
       { OnDisplay: show },
       {
-        where: { id: fileId, DisplayID: displayId, OnDisplay: !show }
+        where: { id: fileId, DisplayID: displayId, OnDisplay: !show },
       }
-
     );
     if (count === 0) return res.json({ success: false });
 
@@ -242,8 +274,8 @@ Files.put("/shown", RoutesCommon.IsAuthenticated, async (req, res) => {
     return res.json({ success: true });
   } catch (err) {
     console.error(err);
-    return res.json({ success: false });
   }
+  return res.json({ success: false });
 });
 
 Files.get("/thumbnail", RoutesCommon.IsAuthenticated, async (req, res) => {
@@ -253,13 +285,14 @@ Files.get("/thumbnail", RoutesCommon.IsAuthenticated, async (req, res) => {
     const displayId = Number(params.id);
 
     const file = await Models.Files.findOne({
-      where: { id: fileId, DisplayID: displayId }
+      where: { id: fileId, DisplayID: displayId },
     });
     if (!file) return res.sendStatus(404);
 
     if (!(await Models.Mongo.File.Exists(file.ThumbnailName)))
       return res.sendStatus(404);
     await Models.Mongo.File.Download(res, file.ThumbnailName);
+    return;
   } catch (err) {
     console.error(err);
     return res.sendStatus(404);
@@ -275,12 +308,16 @@ Files.post(
 
     const files = await Models.Files.findAll({
       attributes: ["id", "Extension", "FileHash"],
-      where: { DisplayID: displayId, Downloaded: false }
+      where: { DisplayID: displayId, Downloaded: false },
     });
 
     const list: any[] = [];
     for (const file of files)
-      list.push({ id: file.id, Name: file.FileHash, Extension: file.Extension });
+      list.push({
+        id: file.id,
+        Name: file.FileHash,
+        Extension: file.Extension,
+      });
 
     return res.json({ success: true, data: list });
   }
@@ -295,13 +332,14 @@ Files.post(
     const displayId = Number(params.id);
 
     const file = await Models.Files.findOne({
-      where: { id: fileId, DisplayID: displayId, Downloaded: false }
+      where: { id: fileId, DisplayID: displayId, Downloaded: false },
     });
 
     if (!file) return res.sendStatus(404);
     if (!(await Models.Mongo.File.Exists(file.FileName)))
       return res.sendStatus(404);
     await Models.Mongo.File.Download(res, file.FileName);
+    return;
   }
 );
 
@@ -315,14 +353,14 @@ Files.post(
       if (params == null)
         return res.json({
           success: false,
-          data: []
+          data: [],
         });
       const id = Number(params.id);
 
       const data: any[] = [];
       const files = await Models.Files.findAll({
         where: { DisplayID: id },
-        order: [["id", "ASC"]]
+        order: [["id", "ASC"]],
       });
 
       for (const file of files)
@@ -332,7 +370,7 @@ Files.post(
           Start: file.TimeStart,
           End: file.TimeEnd,
           ShowTime: file.ShowTime,
-          OnDisplay: file.OnDisplay
+          OnDisplay: file.OnDisplay,
         });
 
       return res.json({ success: true, data: data });
@@ -341,7 +379,7 @@ Files.post(
     }
     return res.json({
       success: false,
-      data: []
+      data: [],
     });
   }
 );
